(feature): initial commit homepage
This commit adds homepage support to mohkale.io. At the moment the
homepage displays the past few posts (as specified in config.yml) and
some highlighted posts. Each post can also now have an associated
thumbnail which is shown on the homepage alongside the post and in the
post itself (just after the title). At the moment the only highlighted
post is the Cheatsheet and it has a basic thumbnail.

There's also a snippet of the day section which justs selects the
solution to a random exercise (that I finished in the past) and
displays it.