(blog): initial commit about page
Quite a lot of changes for such a small addition :cry:. Anyways, now
the about section of my blog actually has some content. Viola. At
the moment it includes a summary section (stating when I was born,
where I'm studying etc.), a skills section which is just a bunch of
languages/frameworks/tools with associated tags related to what I
know, and a timeline which at the moment just says I'm born. I'll
need to fill it out later, but for now this all looks alright.