(CI/CD): Fix build pipelines
The theme upgrade and switch to a submodule structure has lead to the
build pipelines failing. This patch tries to get around that, firstly
a minor annoyance seems to be that the .git submodules directory is only
available in the first task of the build pipeline. If you try to sync
and update submodules in subsequent tasks it'll crash and fail :/.

Secondly I've avoided using the builtin submodule clone features because
one of my submodules is actually private and I didn't want to make it
public or hard-code an access key to force it to be cloned properly.