(scrollback): Move to new toolbar control
So [[orgit-rev:~/prog/repos/mohkale.io/::34895c9][before]] I added a scroll-back button that appears when you scroll past
a certain percentage of the page that when clicked drags you back to the
top of the page (smoothly).

Now I've refactored all that code so you can layer as many such controls
as you require in the ~toolbar~ element. At the moment it should be
pretty much identical to before.